+++ src/vonage/client.py
@@ -13,6 +13,7 @@ from .sms import Sms
 from .ussd import Ussd
 from .voice import Voice
 from .verify import Verify
+from .verify2 import Verify2
 
 import logging
 from platform import python_version
@@ -124,6 +125,7 @@ class Client:
         self.sms = Sms(self)
         self.ussd = Ussd(self)
         self.verify = Verify(self)
+        self.verify2 = Verify2(self)
         self.voice = Voice(self)
 
         self.timeout = timeout
@@ -279,7 +281,11 @@ class Client:
             return None
         elif 200 <= response.status_code < 300:
             # Strip off any encoding from the content-type header:
-            content_mime = response.headers.get("content-type").split(";", 1)[0]
+            try:
+                content_mime = response.headers.get("content-type").split(";", 1)[0]
+            except AttributeError:
+                if response.json() is None:
+                    return None
             if content_mime == "application/json":
                 return response.json()
             else:
@@ -296,7 +302,8 @@ class Client:
                     detail = error_data["detail"]
                     type = error_data["type"]
                     message = f"{title}: {detail} ({type})"
-
+                else:
+                    message = error_data
             except JSONDecodeError:
                 pass
             raise ClientError(message)